functions/requirements.txt
azure-functions
azure-storage-blob
pymongo
cerberus
python-jose
